Вынес функцию throttle  в отдельный модуль util.js
